fix: like moved to a separate component
